[FIX] ddmrp: readd procure_location argument lost in migration
route argument is removed becouse is not used anywhere.

This was lost in migratino to v 16.0.

Original commit details:

[FIX] ddmrp: Check sublocations in Stock Buffers

Sometimes, we want to define a Stock Buffer with entire warehouse visibility but we have Stock Demand Estimates or BoMs defined in sublocations.
This commit allows to see objects defined in sublocations.

Author:    BernatPForgeFlow <bernat.puig@forgeflow.com>
Date:      Mon Jun 5 11:21:34 2023 +0200